+++ extensions/CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Changed
+
+- Group projects are included in `PhylumApi::getProjects` with no group specified
+
 ## 6.4.0 - 2024-05-28
 
 ### Changed