+++ extensions/CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Changed
+
+- Return type fields of `PhylumApi::getPackageDetails`
+
 ## 7.0.0 - 2024-09-17
 
 ### Added