+++ extensions/CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+## 7.1.2 - 2024-10-11
+
 ## 7.1.1 - 2024-10-09
 
 ### Changed